move ticket modal to boardcolumns rename ticket detail to ticket modal
