+++ src/components/Kanban/TicketModal/EditableText.tsx
