+++ src/components/Kanban/index.ts
@@ -1,5 +1,4 @@
 export * from './CreateTickets/CreateTicket'
 export * from './Components/SelectStatus'
-export * from './TicketDetail/TicketDetail'
-export * from './TicketDetail/TicketDetailFunctions'
+export * from './TicketModal/TicketModal'
 export * from './TaskBoard/TaskBoard'
